+++ modules/tidb/src/main/resources/META-INF/services/org.testcontainers.containers.JdbcDatabaseContainerProvider
@@ -1,1 +1,1 @@
-org.testcontainers.containers.TiDBContainerProvider
+org.testcontainers.tidb.TiDBContainerProvider
